+++ sample/src/main/java/me/xiaopan/android/spear/sample/activity/DownloadActivity.java
@@ -170,7 +170,7 @@ public class DownloadActivity extends ActionBarActivity {
                     }
                 }).options(downloadOptions).progressListener(new ProgressListener() {
                     @Override
-                    public void onUpdateProgress(final long totalLength, final long completedLength) {
+                    public void onUpdateProgress(final int totalLength, final int completedLength) {
                         runOnUiThread(new Runnable() {
                             @Override
                             public void run() {
