+++ sample/src/main/java/me/xiaopan/android/spear/sample/adapter/ListImageAdapter.java
@@ -90,7 +90,7 @@ public class ListImageAdapter extends BaseAdapter {
             });
             viewHolder.image.setProgressListener(new ProgressListener() {
                 @Override
-                public void onUpdateProgress(long totalLength, long completedLength) {
+                public void onUpdateProgress(int totalLength, int completedLength) {
                     viewHolder.progressPieView.setProgress((int) (((float) completedLength / totalLength) * 100));
                 }
             });
