+++ sample/src/main/java/me/xiaopan/android/spear/sample/fragment/ImageFragment.java
@@ -72,7 +72,7 @@ public class ImageFragment extends Fragment {
             });
             imageView.setProgressListener(new ProgressListener() {
                 @Override
-                public void onUpdateProgress(long totalLength, long completedLength) {
+                public void onUpdateProgress(int totalLength, int completedLength) {
                     progressBar.setProgress((int) (((float) completedLength / totalLength) * 100));
                 }
             });
